mechanical: move api/fixtures to examples/fixtures
This is support code that is closely tied to the cli, and makes it
difficult to export the api package cleanly.
