+++ cmd/cluster/core/create.go
@@ -21,11 +21,11 @@ import (
 	kubeclient "k8s.io/client-go/kubernetes"
 	crclient "sigs.k8s.io/controller-runtime/pkg/client"
 
-	apifixtures "github.com/openshift/hypershift/api/fixtures"
 	hyperv1 "github.com/openshift/hypershift/api/v1alpha1"
 	"github.com/openshift/hypershift/cmd/log"
 	"github.com/openshift/hypershift/cmd/util"
 	"github.com/openshift/hypershift/cmd/version"
+	apifixtures "github.com/openshift/hypershift/examples/fixtures"
 	apisupport "github.com/openshift/hypershift/support/api"
 )
