+++ cmd/cluster/kubevirt/create.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/spf13/cobra"
 
-	apifixtures "github.com/openshift/hypershift/api/fixtures"
 	"github.com/openshift/hypershift/cmd/cluster/core"
 	"github.com/openshift/hypershift/cmd/log"
+	apifixtures "github.com/openshift/hypershift/examples/fixtures"
 )
 
 const (
