+++ cmd/cluster/none/create.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/spf13/cobra"
 
-	apifixtures "github.com/openshift/hypershift/api/fixtures"
 	"github.com/openshift/hypershift/cmd/cluster/core"
 	"github.com/openshift/hypershift/cmd/log"
+	apifixtures "github.com/openshift/hypershift/examples/fixtures"
 )
 
 func NewCreateCommand(opts *core.CreateOptions) *cobra.Command {
