+++ examples/fixtures/example.go
